"process" decorator now sets LD_LIBRARY_PATH
When the "process" decorator is used, it will call "setup_matlab_runtime()" to set the LD_LIBRARY_PATH and prevent issues on Linux due to MATLAB and PyQt5 library conflicts. Functions no longer need to call "setup_matlab_runtime()" manually, provided that the function created by the process was decorated with "@process".