Get size of repo using GitHub API instead of 'Content-Length' header
Because GitHub uses chunked encoding when downloading the zip file, the 'Content-Length' header does not exist and the file size cannot be used for the progress bar. Instead, we make an API request to retrieve the size of the repository (which is an estimate for the size of the zip file), and pass the size to DownloadThread.