Fix exception handling on Windows
On Windows, PyQt suppresses exceptions and crashes with a generic error. This makes developing difficult on Windows, but it can be fixed by manually calling the system exception hook using a custom exception hook (see https://stackoverflow.com/a/37837374/6364450).

Since PyMODA already uses a custom exception hook, it was simple to modify its behaviour.

The new behaviour is as follows:
- In normal mode, PyMODA shows a dialog and the system exception hook prints a traceback.
- In debug mode, PyMODA calls the system exception hook to print a traceback. No dialog is shown.