Return dict with 'fmin' from WT; save params including 'fmin'
The parameters used to perform the calculation should be saved to the data file when using 'save data'; however, if not specified in the GUI, the value of 'fmin' is calculated by the WT at runtime. This meant that it was not possible to save the value of 'fmin'.

To solve this problem, the `wt.m` file was updated to take an optional boolean parameter `python`, which defaults to false. PyMODA supplies `python=True`. When the `python` parameter is True, the MATLAB function returns the WT, frequencies and a dictionary containing only values which can be safely returned (no structs etc).