Use 'run_in_thread=True' when creating Scheduler on Windows/macOS
To improve performance when performing a single calculation, PyMODA now takes advantage of a new AsyncProcessScheduler feature which allows the first task to be run in a thread instead of a process. This is important on Windows/macOS, since they suffer much larger overhead due to spawning processes.

Linux will not use this feature, because:
1. Forking processes is fast.
2. Calling MATLAB functions on a thread would cause issues with the
   LD_LIBRARY_PATH.

The delay due to pre-processing is no longer present, and performing a
wavelet transform, etc, on a single transform is faster.